+++ OvmfPkg/Include/IndustryStandard/QemuLoader.h
